requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-countries==6.1.2
django-stdimage==5.1.1
future==0.18.2
kombu==4.6.11
Pillow==7.2.0
progressbar2==3.51.4
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
